+++ sim/paladin/paladin.go
@@ -25,6 +25,16 @@ const (
 	SpellCode_PaladinConsecration
 )
 
+type SealJudgeCode uint8
+
+const (
+	SealJudgeCodeNone        SealJudgeCode = 0
+	SealJudgeOfMartyrdomCode SealJudgeCode = 1 << iota
+	SealJudgeOfRighteousnessCode
+	SealJudgeOfCommandCode
+	SealJudgeOfTheCrusaderCode
+)
+
 type Paladin struct {
 	core.Character
 
@@ -36,16 +46,19 @@ type Paladin struct {
 	currentPaladinAura *core.Aura
 
 	currentSeal      *core.Aura
-	auraSoM          *core.Aura
-	aurasSoR         [8]*core.Aura
-	aurasSoC         [5]*core.Aura
-	aurasSotC        [6]*core.Aura
+	allSealAuras     [][]*core.Aura
+	aurasSoM         []*core.Aura
+	aurasSoR         []*core.Aura
+	aurasSoC         []*core.Aura
+	aurasSotC        []*core.Aura
 	currentJudgement *core.Spell
 	allJudgeSpells   [][]*core.Spell
 	spellsJoM        []*core.Spell
 	spellsJoR        []*core.Spell
 	spellsJoC        []*core.Spell
 	spellsJotC       []*core.Spell
+
+	rollDummyJudgeHit [4]bool
 
 	// Active abilities and shared cooldowns that are externally manipulated.
 	holyShockCooldown *core.Cooldown
@@ -61,15 +74,12 @@ type Paladin struct {
 	sealOfCommand       *core.Spell
 	sealOfMartyrdom     *core.Spell
 
-	lingerDuration time.Duration
+	lingerDuration      time.Duration
+	consumeSealsOnJudge bool
 
 	// T2 Bonuses Related (Draconic)
-	t2Judgement2pc        bool
-	t2Judgement2pcApplied bool
-	t2Judgement4pc        bool
-	t2Judgement6pc        bool
-	t2Judgement6pcAura    *core.Aura
-	lastJudgement         int
+	thisJudgement SealJudgeCode
+	lastJudgement SealJudgeCode
 }
 
 // Implemented by each Paladin spec.
@@ -106,6 +116,13 @@ func (paladin *Paladin) Initialize() {
 	paladin.allJudgeSpells = append(paladin.allJudgeSpells, paladin.spellsJoC)
 	paladin.allJudgeSpells = append(paladin.allJudgeSpells, paladin.spellsJotC)
 
+	paladin.allSealAuras = append(paladin.allSealAuras, paladin.aurasSoM)
+	paladin.allSealAuras = append(paladin.allSealAuras, paladin.aurasSoR)
+	paladin.allSealAuras = append(paladin.allSealAuras, paladin.aurasSoC)
+	paladin.allSealAuras = append(paladin.allSealAuras, paladin.aurasSotC)
+
+	paladin.rollDummyJudgeHit = [4]bool{false, false, true, false}
+
 	// Active abilities
 	paladin.registerCrusaderStrike()
 	paladin.registerDivineStorm()
@@ -119,18 +136,12 @@ func (paladin *Paladin) Initialize() {
 	paladin.registerAuraMastery()
 
 	paladin.lingerDuration = time.Millisecond * 400
+	paladin.consumeSealsOnJudge = true
 
 	paladin.registerStopAttackMacros()
-
-	paladin.t2Judgement2pc = false
-	paladin.t2Judgement2pcApplied = false
-	paladin.t2Judgement4pc = false
-	paladin.t2Judgement6pc = false
-
 }
 
 func (paladin *Paladin) Reset(_ *core.Simulation) {
-	//paladin.registerStopAttackMacros()
 }
 
 // maybe need to add stat dependencies
