+++ sim/paladin/soc.go
@@ -126,7 +126,7 @@ func (paladin *Paladin) registerSealOfCommand() {
 			},
 		})
 
-		paladin.aurasSoC[i] = aura
+		paladin.aurasSoC = append(paladin.aurasSoC, aura)
 
 		paladin.sealOfCommand = paladin.RegisterSpell(core.SpellConfig{
 			ActionID:    aura.ActionID,
